10 rate version based

+++ src/main/java/ir/niopdc/policy/domain/policy/PolicyRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface PolicyRepository extends ListCrudRepository<Policy, Integer> {
+interface PolicyRepository extends ListCrudRepository<Policy, Byte> {
 }
